Make force_mda in the esmtprc to work with the -t option (based on patch by Sourav K. Mandal).
force_mda in the esmtprc does not work with the -t option passed to
esmtp, because rcfile_parse is called after message_parse_headers is
called in the -t handling.  Of course, message_parse_headers relies on
the esmtprc to have been parsed already in order to denote all
recipients as local when force_mda is specified.